Update tutorial-11.md
It seems domina.events has been refactored to remove the builtin-events variable.

"issues #27 & #63 - removed the use of built-in event types. All keywords are now just converted directly."

https://github.com/levand/domina/commit/6960d0b9b9d63fa71f06d0ee7203becf99e00294

This change provides an alternative for how the available types can be inspected.  It's getting a little complicated.  An alternative might be to remove this section.